tests/wpt/tests/html/semantics/the-button-element/interest-target/interesttarget-pseudo-classes.tentative.html

<!DOCTYPE html>
<meta charset="utf-8" />
<link rel="author" href="mailto:masonf@chromium.org">
<link rel="help" href="https://open-ui.org/components/interest-invokers.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/invoker-utils.js"></script>

<div id=unrelated tabindex=0>Unrelated</div>
<button id=invoker interesttarget=target>Invoker</button>
<div id=target>Target <button id=target_button>contained button</button></div>
<style>
  button {
    interest-target-delay: 0s;
  }
</style>
<script>
function checkPseudos(invoker,target,expectHasPartialInterest,expectHasInterest,expectTargetHasPartialInterest,expectTargetHasInterest,msg) {
  msg = msg ?? 'Error';
  assert_true(!expectHasPartialInterest || expectHasInterest,'Partial interest can\'t be true without full interest');
  assert_true(!expectTargetHasPartialInterest || expectTargetHasInterest,'Partial interest can\'t be true without full interest (target)');
  assert_equals(invoker.matches(':has-interest'),expectHasInterest,`${msg}: :has-interest mismatch`);
  assert_equals(invoker.matches(':has-partial-interest'),expectHasPartialInterest,`${msg}: :has-partial-interest mismatch`);
  assert_equals(target.matches(':target-of-interest'),expectTargetHasInterest,`${msg}: :target-of-partial-interest mismatch`);
  assert_equals(target.matches(':target-of-partial-interest'),expectTargetHasPartialInterest,`${msg}: :target-of-partial-interest mismatch`);
  assert_false(invoker.matches(':target-of-interest'),'invoker should never match :target-of-interest');
  assert_false(invoker.matches(':target-of-partial-interest'),'invoker should never match :target-of-partial-interest');
  assert_false(target.matches(':has-interest'),'target should never match :has-interest');
  assert_false(target.matches(':has-partial-interest'),'target should never match :has-partial-interest');
}
function add_cleanups(t) {
  t.add_cleanup(async() => {
    invoker.removeAttribute('style');
    await focusOn(unrelated);
    await hoverOver(unrelated);
  });
}

promise_test(async (t) => {
  add_cleanups(t);
  let hasInterest = false;
  target.addEventListener('interest',() => (hasInterest=true));
  target.addEventListener('loseinterest',() => (hasInterest=false));
  checkPseudos(invoker,target,false,false,false,false,'initial');
  assert_false(hasInterest);
  await hoverOver(invoker);
  checkPseudos(invoker,target,false,true,false,true,'hovering invoker shows full interest (and not partial interest)');
  assert_true(hasInterest,'event was fired');
  await hoverOver(target);
  checkPseudos(invoker,target,false,true,false,true,'hovering the target maintains interest');
  assert_true(hasInterest,'loseinterest event was not yet fired');
  await hoverOver(unrelated);
  checkPseudos(invoker,target,false,false,false,false,'hovering unrelated loses interest');
  assert_false(hasInterest,'loseinterest event was fired');
},'Basic pseudo class function, with mouse hover triggering');

promise_test(async (t) => {
  add_cleanups(t);
  let hasInterest = false;
  target.addEventListener('interest',() => (hasInterest=true));
  target.addEventListener('loseinterest',() => (hasInterest=false));
  checkPseudos(invoker,target,false,false,false,false,'initial');
  assert_false(hasInterest);
  await focusOn(invoker);
  checkPseudos(invoker,target,true,true,true,true,'focusing invoker shows partial interest (and therefore also interest)');
  assert_true(hasInterest,'event was fired');
  await focusOn(target_button);
  checkPseudos(invoker,target,false,true,false,true,'focusing the target upgrades to full interest (no longer partial interest)');
  assert_true(hasInterest,'loseinterest event was not yet fired');
  await focusOn(invoker);
  checkPseudos(invoker,target,false,true,false,true,'focusing back on invoker keeps full interest');
  assert_true(hasInterest,'loseinterest event was not yet fired');
  await focusOn(unrelated);
  checkPseudos(invoker,target,false,false,false,false,'focusing unrelated loses interest');
  assert_false(hasInterest,'loseinterest event was fired');
},'Basic pseudo class function, with keyboard focus triggering');

const invokerDelayMs = 100; // The CSS delay setting.
const hoverWaitTime = 200; // How long to wait to cover the delay for sure.
promise_test(async (t) => {
  add_cleanups(t);
  invoker.setAttribute('style',`interest-target-delay: ${invokerDelayMs}ms`);
  checkPseudos(invoker,target,false,false,false,false,'initial');
  const token1 = await mouseOverAndRecord(t,invoker);
  const immediate_result = invoker.matches(':has-interest') ||
      invoker.matches(':has-partial-interest') ||
      target.matches(':target-of-interest') ||
      target.matches(':target-of-partial-interest');
  if (msSinceMouseOver(token1) < invokerDelayMs) {
    assert_false(immediate_result,'No pseudos should match before the show delay elapses');
  }
  await waitForHoverTime(hoverWaitTime);
  checkPseudos(invoker,target,false,true,false,true,'full interest pseudos (but not partial interest) should match after hover delay');
  const token2 = await mouseOverAndRecord(t,unrelated);
  const immediate_result2 =  invoker.matches(':has-interest') &&
      !invoker.matches(':has-partial-interest') &&
      target.matches(':target-of-interest') &&
      !target.matches(':target-of-partial-interest');
  if (msSinceMouseOver(token2) < invokerDelayMs) {
    assert_true(immediate_result2,'all pseudos should still represent full interest before the hide delay elapses');
  }
  await waitForHoverTime(hoverWaitTime);
  checkPseudos(invoker,target,false,false,false,false,'no pseudos should match after de-hover delay');
},'The pseudo classes only match after delays, once interest is shown');
</script>
